% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_communality.R
\name{get_communality}
\alias{get_communality}
\title{get_communality}
\usage{
get_communality(mv_name, lv_name, tau, data)
}
\arguments{
\item{mv_name}{manifest variable label (string)}

\item{lv_name}{latent variable label (string)}

\item{tau}{the quantile(s) to be estimated}

\item{data}{dataset. It includes manifest variable and latent score 
for a specific quantile}
}
\value{
the communality for each manifest varaible
}
\description{
get_communality
}
\details{
Internal function. \code{get_communality} is called by \code{assessment} and
}
\keyword{internal}
