% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_element.R
\name{get_element}
\alias{get_element}
\title{get_element}
\usage{
get_element(path_matrix)
}
\arguments{
\item{path_matrix}{the matrix of path coefficients}
}
\value{
the path coefficients labels
}
\description{
get_element
}
\details{
Internal function. \code{get_element} is called by \code{get_paths}
}
\keyword{internal}
