% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_internal_parameter_estimation.R
\name{get_internal_parameter_estimation}
\alias{get_internal_parameter_estimation}
\title{get_internal_parameter_estimation}
\usage{
get_internal_parameter_estimation(
  X,
  modes,
  tau,
  lvs,
  lvs.names,
  IDM,
  sets,
  scheme,
  method,
  fix.quantile,
  tol,
  maxiter
)
}
\arguments{
\item{X}{matrix or data frame containing the manifest variables.}

\item{modes}{character vector indicating the type of measurement for each
block. Possible values are: \code{"A", "B"}. 
The length of \code{modes} must be equal to the length of \code{blocks}.}

\item{tau}{if sepcifed indicate the specific quantile to be considered}

\item{lvs}{the number of latent variables}

\item{lvs.names}{the label of latent variables}

\item{IDM}{the path matrix}

\item{sets}{the outer model}

\item{scheme}{the internal scheme}

\item{method}{rq method. It is equal to \code{"br"} by default}

\item{fix.quantile}{is boolean equal to \code{FALSE} or \code{TRUE} indicating 
whether invariance is holded.}

\item{tol}{decimal value indicating the tolerance criterion for the iterations (tol=0.00001).}

\item{maxiter}{integer indicating the maximum number of iterations (maxiter=100 by default).}
}
\value{
the outer weights
}
\description{
get_internal_parameter_estimation
}
\details{
Internal function. \code{get_internal_parameter_estimation} is called by \code{qcpm} and
}
\keyword{internal}
