% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_loadings.R
\name{get_loadings}
\alias{get_loadings}
\title{get_loadings}
\usage{
get_loadings(data, sets, mvs, lvs, IDM, tau, LV, qcorr, ...)
}
\arguments{
\item{data}{the matrix of data (manifest variables)}

\item{sets}{outer model}

\item{mvs}{number of manifest variables}

\item{lvs}{number of latent variables}

\item{IDM}{the path matrix}

\item{tau}{the quantile(s) to be estimated}

\item{LV}{the  estimated latent variables}

\item{qcorr}{boolean. If it si equal to \code{TRUE},
loadings are standardized.}
}
\value{
the loadings estimated for each latent variables
}
\description{
get_loadings
}
\details{
Internal function. \code{get_loadings} is called by \code{qcpm} and
}
\keyword{internal}
