% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paths.R
\name{get_paths}
\alias{get_paths}
\title{get_paths}
\usage{
get_paths(path_matrix, Y_lvs, tau, full = TRUE, ...)
}
\arguments{
\item{path_matrix}{the matrix of path coefficients}

\item{Y_lvs}{the matrix of latent variables}

\item{tau}{the quantile(s) to be estimated}
}
\value{
the path coefficients
}
\description{
get_paths
}
\details{
Internal function. \code{get_paths} is called by \code{qcpm} and
}
\keyword{internal}
