% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qc.R
\name{qc}
\alias{qc}
\title{qc}
\usage{
qc(x, y, tau)
}
\arguments{
\item{x}{the covariate variables}

\item{y}{the dependent variable}

\item{tau}{the quantile(s) to be estimated}
}
\value{
the qc correlation
}
\description{
qc
}
\details{
Internal function. \code{qc} is called by \code{get_internal_parameter_estimation} and
}
\keyword{internal}
